requirements.txt
# Core Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
redis==5.0.1
django-redis==5.4.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
cryptography==41.0.7

# Machine Learning & Analytics
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.25.2
tensorflow==2.15.0
torch>=2.0.0
xgboost>=1.7.0
joblib==1.3.2

# API Documentation
drf-spectacular==0.26.5
drf-spectacular[sidecar]==0.26.5

# Monitoring & Logging
celery==5.3.4
django-celery-beat==2.5.0
sentry-sdk==1.38.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
Pillow==10.1.0

# Development
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0